pyproject.toml
[project]
name = "PDFtoMD"
version = "3.0.0"
description = "Enterprise-ready PDF to Markdown conversion tool powered by multimodal AI. Features secure SSOT configuration, centralized API key management, comprehensive security controls, and enterprise-grade batch processing capabilities."
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Joseph Wright", email = "joey.wright@outlook.com"},
]
maintainers = [
    {name = "Joseph Wright", email = "joey.wright@outlook.com"},
]
keywords = ["pdf", "markdown", "ai", "llm", "conversion", "batch-processing", "openai", "enterprise", "security", "ssot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
    "openai>=1.97.0",
    "pymupdf>=1.26.3",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
]

[project.urls]
Homepage = "https://github.com/ch0t4nk/PDFtoMD"
Repository = "https://github.com/ch0t4nk/PDFtoMD"
Issues = "https://github.com/ch0t4nk/PDFtoMD/issues"
Documentation = "https://github.com/ch0t4nk/PDFtoMD#readme"
"Original Project" = "https://github.com/MarkPDFdown/markpdfdown"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest==8.3.5",
    "ruff>=0.11.11",
]

# Ruff 配置
[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "UP007", # Use `X | Y` for type annotations
    "UP036", # Version comparisons (needed for backwards compatibility)
    "E402",  # Module level import not at top (needed for dynamic path additions)
    "F841",  # Local variable assigned but never used (acceptable in some contexts)
    "UP038", # Use `X | Y` in isinstance (can be less readable in some cases)
    "C401",  # Unnecessary generator (sometimes more readable)
    "F541",  # f-string without placeholders (acceptable for consistency)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["markpdfdown"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["markpdfdown"]

[tool.hatch.build.targets.wheel.force-include]
"." = "."
